+++ spec/kettle/soup/cover/loaders_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe Kettle::Soup::Cover::Loaders do
 
   before do
     @original_formatters = SimpleCov.formatters
-    Kettle::Soup::Cover::Constants.reset_const do
+    reset_soup_cover_constants do
       stub_env("CI" => ci)
       stub_env("K_SOUP_COV_FORMATTERS" => formatters)
     end